Complete Region Overview(InnerSub Component) -- Prajjwol
